+++ Runtime/Theme/Swatch/ESwatchType.cs
@@ -1,0 +1,9 @@
+namespace Theme.Swatch
+{
+  public enum ESwatchType
+  {
+    None = 0,
+    Color = 1,
+    Gradient = 2
+  }
+}